Move file for conversion functions since it makes much more sense to keep the conversion functions close to the actual parser Also in this commit: outxml_parser now accepts a file handle as input
